+++ models/db/Motion.php
@@ -409,6 +409,15 @@ class Motion extends IMotion implements IRSSItem
         return $intro . $this->title;
     }
 
+    public function showTitlePrefix(): bool
+    {
+        return (
+            !$this->getMyConsultation()->getSettings()->hideTitlePrefix &&
+            trim($this->getFormattedTitlePrefix()) !== '' &&
+            !$this->isResolution()
+        );
+    }
+
     public function getTitleWithPrefix(): string
     {
         if ($this->getMyConsultation()->getSettings()->hideTitlePrefix) {