+++ models/db/User.php
@@ -4,11 +4,7 @@ namespace app\models\db;
 
 use app\models\layoutHooks\Layout;
 use app\models\settings\PrivilegeQueryContext;
-use app\components\{ExternalPasswordAuthenticatorInterface,
-    RequestContext,
-    Tools,
-    UrlHelper,
-    mail\Tools as MailTools};
+use app\components\{ExternalPasswordAuthenticatorInterface, MotionNumbering, RequestContext, Tools, UrlHelper, mail\Tools as MailTools};
 use app\models\events\UserEvent;
 use app\models\exceptions\{ExceptionBase, FormError, MailNotSent, ServerConfiguration};
 use app\models\settings\AntragsgruenApp;
@@ -523,19 +519,35 @@ class User extends ActiveRecord implements IdentityInterface
      */
     public function getMySupportedMotionsByConsultation(Consultation $consultation): array
     {
-        $query     = MotionSupporter::find();
+        $query = MotionSupporter::find();
         $query->innerJoin(
             'motion',
             'motionSupporter.motionId = motion.id'
         );
-        $query->where('motion.status != ' . IntVal(Motion::STATUS_DELETED));
-        $query->andWhere('motion.consultationId = ' . IntVal($consultation->id));
-        $query->andWhere('motionSupporter.userId = ' . IntVal($this->id));
-        $query->orderBy('(motionSupporter.role = "initiates") DESC, motion.dateCreation DESC');
+        $query->where('motion.status != ' . intval(Motion::STATUS_DELETED));
+        $query->andWhere('motion.consultationId = ' . intval($consultation->id));
+        $query->andWhere('motionSupporter.userId = ' . intval($this->id));
+        $query->orderBy('(motionSupporter.role = "initiates") DESC, motion.titlePrefix ASC, motion.dateCreation DESC, motion.id DESC');
 
         /** @var MotionSupporter[] $supporters */
         $supporters = $query->all();
-        return $supporters;
+
+        // Hint: we go through the supports, from the newest motion version to the oldest, and keep track of the root motions already seen.
+        // Skipping older entries resolving to the same root motion. Thus, we will only return the most recent version of each motion.
+        $filteredSupporters = [];
+        $firstMotionIds = [];
+        foreach ($supporters as $supporter) {
+            /** @var Motion $motion */
+            $motion = $supporter->getIMotion();
+            $history = MotionNumbering::getSortedHistoryForMotion($motion, false);
+            if (count($history) === 0 || in_array($history[0]->id, $firstMotionIds)) {
+                continue;
+            }
+            $firstMotionIds[] = $history[0]->id;
+            $filteredSupporters[] = $supporter;
+        }
+
+        return $filteredSupporters;
     }
 
     /**
@@ -553,7 +565,7 @@ class User extends ActiveRecord implements IdentityInterface
         $query->andWhere('amendment.status != ' . IntVal(Motion::STATUS_DELETED));
         $query->andWhere('motion.consultationId = ' . IntVal($consultation->id));
         $query->andWhere('amendmentSupporter.userId = ' . IntVal($this->id));
-        $query->orderBy('(amendmentSupporter.role = "initiates") DESC, amendment.dateCreation DESC');
+        $query->orderBy('(amendmentSupporter.role = "initiates") DESC, motion.titlePrefix ASC, amendment.titlePrefix ASC, amendment.dateCreation DESC');
         /** @var AmendmentSupporter[] $supporters */
         $supporters = $query->all();
         return $supporters;